+++ src/admin-orderable-edit/admin-orderable-edit.routes.spec.js
@@ -145,14 +145,14 @@ describe('openlmis.administration.orderables.edit route', function() {
 
     describe('.programs state', function() {
 
-        it('should resolve orderable', function() {
+        it('should resolve programs orderable', function() {
             this.goToState('/programs');
 
-            expect(this.getResolvedValue('orderable')).toEqual(this.orderable);
-            expect(this.getResolvedValue('orderable')).not.toBe(this.orderable);
+            expect(this.getResolvedValue('programsOrderable')).toEqual(this.orderable.programs);
+            expect(this.getResolvedValue('programsOrderable')).not.toBe(this.orderable.programs);
         });
 
-        it('should not change state if fetching orderable fails', function() {
+        it('should not change state if fetching programsOrderable fails', function() {
             this.OrderableResource.prototype.query.andReturn(this.$q.reject());
 
             this.goToState('/programs');
@@ -169,6 +169,36 @@ describe('openlmis.administration.orderables.edit route', function() {
             expected[this.programs[2].id] = this.programs[2];
 
             expect(this.getResolvedValue('programsMap')).toEqual(expected);
+        });
+
+        it('should resolve canEdit as true if user has right to edit programs', function() {
+
+            this.goToState('/programs');
+
+            expect(this.getResolvedValue('canEdit')).toBeTruthy();
+
+        });
+
+        it('should resolve canEdit as false if user does not have right to edit program orderable', function() {
+            var context = this;
+            this.permissionService.hasPermissionWithAnyProgramAndAnyFacility.andCallFake(function() {
+                return context.$q.reject();
+            });
+            this.goToState('/programs');
+
+            expect(this.getResolvedValue('canEdit')).toBeFalsy();
+        });
+
+        it('should resolve canEdit as true if user have right to edit program orderable', function() {
+            var context = this;
+            this.permissionService.hasPermissionWithAnyProgramAndAnyFacility.andCallFake(function(userId, params) {
+                return params.right === context.ADMINISTRATION_RIGHTS.ORDERABLES_MANAGE
+                    ? context.$q.resolve() : context.$q.reject();
+            });
+            this.goToState('/programs');
+
+            expect(this.getResolvedValue('canEdit')).toBeTruthy();
+
         });
 
         it('should not change state if fetching programs fails', function() {